geotiff tests: consolidate input-validation cluster (#2430)
Fold nine top-level input-validation test files into a new
xrspatial/geotiff/tests/unit/test_input_validation.py, organised by
validation axis: band type/bool, size params, source dimensions, 3D
writer dims, window bounds, and degenerate pixel size.

Update the reader.windowed release-gate row to cite the new file so
the checklist-parity gate stays green.
